+++ docker/docker-entrypoint.sh
@@ -70,8 +70,9 @@ if [[ "$1" == wsgi ]]; then
 elif [[ "$1" == celeryd ]]; then
   cd $APP_DIR
 
-  exec celery worker \
+  exec celery \
     --app config.celery_app \
+    worker \
     -Q "${CELERY_QUEUES}" \
     --concurrency "${CELERY_WORKERS}" \
     --loglevel info
@@ -79,9 +80,10 @@ elif [[ "$1" == celerybeat ]]; then
   cd $APP_DIR
   rm -f celerybeat.pid
 
-  exec celery beat \
-    --max-interval 30 \
+  exec celery \
     --app config.celery_app \
+    beat \
+    --max-interval 30 \
     --loglevel info
 else
   cd $APP_DIR