+++ CHANGELOG.rst
@@ -27,6 +27,8 @@ Changed
 Fixed
 -----
 
+- **General**
+    - Docker entry points for Celery and Celerybeat (#1193)
 - **Samplesheets**
     - Loading icon in vue app iRODS status badge (#1192)